pyproject.toml
#pyproject v.0.0.0
[tool.poetry]
name = "ma-mapper"
version = "0.8.0"
description = "Multiple Alignment Mapper underconstruction"
authors = ["pkncsk <pkncsk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
biopython = "^1.83"
compress-pickle = "^2.1.0"
cyvcf2 = "^0.30.28"
h5py = "^3.7.0"
logomaker = "^0.8"
matplotlib = "^3.8.4"
numpy = "^1.21.5"
pandas = "^1.3.5"
plotly = "^5.22.0"
pybedtools = "^0.10.0"
pybigwig = "^0.3.23"
pysam = "^0.22.0"
scipy = "^1.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
